+++ submission/persistence/src/main/kotlin/ac/uk/ebi/biostd/persistence/model/User.kt
@@ -60,9 +60,6 @@ class User(
     @Column
     var activationKey: String? = null
 
-    @Column(name = "ssoSubject")
-    var domainGroup: String? = null
-
     @Column(name = "auxProfileInfo")
     @Convert(converter = AuxInfoConverter::class)
     var auxInfo: AuxInfo = AuxInfo()
