--- submission/submission-security/src/main/kotlin/ebi/ac/uk/security/integration/components/IAutomatedSecurityService.kt
@@ -1,7 +1,0 @@
-package ebi.ac.uk.security.integration.components
-
-import ebi.ac.uk.security.integration.model.api.SecurityUser
-
-interface IAutomatedSecurityService {
-    fun getOrCreate(email: String, username: String): SecurityUser
-}
