+++ submission/submission-security/src/main/kotlin/ebi/ac/uk/security/integration/components/ISecurityService.kt
@@ -27,4 +27,6 @@ interface ISecurityService {
     fun resetPassword(request: ResetPasswordRequest)
 
     fun getUserProfile(authToken: String): UserInfo
+
+    fun getOrCreateInactive(email: String, username: String): SecurityUser
 }
